fix(execution): remove killing pending processes
delete function 'finish all' that kill the other process after the last one is done. Also removing 'pid' member from the 'cmd' struct since it was only storing the process id to kill it later.

reason to remove this function is that the other processes should be waited till they finish instead of forcing them to stop, this was a previous fix to stop processes that wait for stdin input from pending and it will replaced  by closing the pipe fd.